__init__以外のスクリプトをscriptsフォルダに格納
