+++ scripts/func_select_axis_from_point.py
