+++ scripts/menu_edit_mesh_context.py
