+++ scripts/translations.py
